+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Try to keep listed changes to a concise bulleted list of simple explanations of changes. Aim for the amount of information needed so that readers can understand where they would look in the codebase to investigate the changes' implementation, or where they would look in the documentation to understand how to make use of the change in practice - better yet, link directly to the docs and provide detailed information there. Only elaborate if doing so is required to avoid breaking changes or experimental features from ruining someone's day.
 
 ## [Unreleased]
+### Added
+- Updated grpc logs for the healthcheck service to log at the Debug level instead of at the Info level. [#2340](https://github.com/openfga/openfga/pull/2340)
 
 ### Added
 - Separate out experimental list objects optimization flag (`enable-list-objects-optimizations`) from experimental check optimization flag (`enable-check-optimizations`) to allow individual optimization. [#2341](https://github.com/openfga/openfga/pull/2341).